0.1 version with unit tests

+++ orphanage/src/main/java/pl/polsl/orphanage/service/RelativeService.java
@@ -2,13 +2,16 @@ package pl.polsl.orphanage.service;
 
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import pl.polsl.orphanage.domain.Fosterling;
 import pl.polsl.orphanage.domain.Relative;
+import pl.polsl.orphanage.repository.FosterlingRepository;
 import pl.polsl.orphanage.repository.RelativeRepository;
 import pl.polsl.orphanage.service.dto.RelativeDTO;
 import pl.polsl.orphanage.service.mapper.RelativeMapper;
 
 import java.util.LinkedList;
 import java.util.List;
+import java.util.Optional;
 import java.util.stream.Collectors;
 
 /**
@@ -22,9 +25,15 @@ public class RelativeService {
 
     private RelativeMapper relativeMapper;
 
-    public RelativeService(RelativeRepository relativeRepository, RelativeMapper relativeMapper) {
+    private FosterlingRepository fosterlingRepository;
+
+    public RelativeService(
+            RelativeRepository relativeRepository,
+            RelativeMapper relativeMapper,
+            FosterlingRepository fosterlingRepository) {
         this.relativeRepository = relativeRepository;
         this.relativeMapper = relativeMapper;
+        this.fosterlingRepository = fosterlingRepository;
     }
 
     /**
@@ -60,7 +69,7 @@ public class RelativeService {
      * @return list of entity
      */
     public List<RelativeDTO> findAllByLastname(String lastname){
-        return relativeRepository.findAllByLastname(lastname).stream()
+        return relativeRepository.findAllByLastname("%"+lastname+"%").stream()
                 .map(relativeMapper::toDto)
                 .collect(Collectors.toCollection(LinkedList::new));
     }
@@ -70,7 +79,45 @@ public class RelativeService {
      *
      * @param id the id of the entity
      */
-    public void delete(Long id){
-        relativeRepository.deleteById(id);
+    public boolean delete(Long id){
+
+        Optional<Relative> relative = relativeRepository.findById(id);
+        if(relative.isPresent()) {
+
+            List<Fosterling> fosterlings = fosterlingRepository.findAllWithRelative(id);
+
+            fosterlings
+                    .stream()
+                    .forEach(
+                            (x) -> x.getRelatives().remove(relative.get())
+                    );
+
+            fosterlingRepository.saveAll(fosterlings);
+
+            relativeRepository.deleteById(id);
+
+            return true;
+        }
+        return false;
+    }
+
+    /**
+     * Add relative to fosterling by id
+     *
+     * @param id the id of relative id
+     * @param fosterlingId the id of fosterling id
+     */
+    public boolean addRelativeToFosterling(Long id, Long fosterlingId){
+        Optional<Fosterling> fosterling = fosterlingRepository.findById(fosterlingId);
+        Optional<Relative> relative = relativeRepository.findById(id);
+        if(fosterling.isPresent() && relative.isPresent()){
+
+            fosterling.get().getRelatives().add(relative.get());
+            fosterlingRepository.save(fosterling.get());
+            return true;
+        }
+
+        return false;
+
     }
 }
